+++ doc/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Better error handling during type conversion when reading lattice files
 * Add option to shift the lattice in periodic directions
+* Add dialog to display information about a loaded system
 
 **150204.1**